# app/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a modern RSS reader API.",
        "title": "Krss API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/api/proxy/image/{encoded}": {
            "get": {
                "description": "Proxies external images to avoid triggering anti-crawling mechanisms",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "proxy"
                ],
                "summary": "Proxy external image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base64 URL-safe encoded image URL",
                        "name": "encoded",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Base64 URL-safe encoded article URL (used as Referer for CDN anti-hotlinking)",
                        "name": "ref",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "504": {
                        "description": "Gateway Timeout",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}